requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
addict=2.4.0=pypi_0
altair=5.1.2=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.6.2=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
black=23.3.0=pypi_0
bleach=6.0.0=pypi_0
blinker=1.6.2=pypi_0
blis=0.7.11=pypi_0
braceexpand=0.1.7=pypi_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.0=pypi_0
catalogue=2.0.10=pypi_0
ccimport=0.4.2=pypi_0
certifi=2023.5.7=pypi_0
cffi=1.15.1=pypi_0
cfgv=3.4.0=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
cloudpathlib=0.15.1=pypi_0
cloudpickle=2.2.1=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.3=pypi_0
confection=0.1.3=pypi_0
contexttimer=0.3.3=pypi_0
contourpy=1.0.7=pypi_0
cumm=0.4.11=pypi_0
cycler=0.11.0=pypi_0
cymem=2.0.8=pypi_0
cython=0.29.34=pypi_0
dask=2023.3.2=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pypi_0
decord=0.6.0=pypi_0
defusedxml=0.7.1=pypi_0
descartes=1.1.0=pypi_0
detectron2=0.6=pypi_0
distlib=0.3.7=pypi_0
distro=1.8.0=pypi_0
dm-tree=0.1.8=pypi_0
docker-pycreds=0.4.0=pypi_0
exceptiongroup=1.1.1=pypi_0
executing=1.2.0=pypi_0
fairscale=0.4.4=pypi_0
fastjsonschema=2.16.3=pypi_0
filelock=3.12.4=pypi_0
fire=0.5.0=pypi_0
flake8=5.0.4=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.39.3=pypi_0
fqdn=1.5.1=pypi_0
fsspec=2023.4.0=pypi_0
ftfy=6.1.1=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
google-auth=2.20.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
googleapis-common-protos=1.59.1=pypi_0
grpcio=1.54.0=pypi_0
h11=0.14.0=pypi_0
h5py=3.8.0=pypi_0
httpcore=1.0.1=pypi_0
httpx=0.25.1=pypi_0
huggingface-hub=0.14.1=pypi_0
hydra-core=1.3.2=pypi_0
identify=2.5.30=pypi_0
idna=3.4=pypi_0
imageio=2.28.1=pypi_0
importlib-metadata=6.6.0=pypi_0
importlib-resources=5.12.0=pypi_0
iniconfig=2.0.0=pypi_0
iopath=0.1.9=pypi_0
ipykernel=6.22.0=pypi_0
ipython=8.12.2=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.6=pypi_0
isoduration=20.11.0=pypi_0
isort=5.13.2=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.2.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.3.0=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.5.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.7=pypi_0
kaggle=1.5.16=pypi_0
kiwisolver=1.4.4=pypi_0
kornia=0.5.0=pypi_0
langcodes=3.3.0=pypi_0
lark=1.1.7=pypi_0
lazy-loader=0.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.0=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.31.0=pypi_0
lyft-dataset-sdk=0.0.8=pypi_0
markdown=3.4.3=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.2.0=pypi_0
matplotlib-inline=0.1.6=pypi_0
mccabe=0.7.0=pypi_0
mdurl=0.1.2=pypi_0
mistune=2.0.5=pypi_0
mmcls=0.25.0=pypi_0
mmcv-full=1.4.0=pypi_0
mmdet=2.22.0=dev_0
mmdet3d=0.17.1=dev_0
mmsegmentation=0.14.1=pypi_0
model-index=0.1.11=pypi_0
murmurhash=1.0.10=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.4=pypi_0
nbconvert=7.3.1=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
networkx=2.2=pypi_0
ninja=1.11.1=pypi_0
nodeenv=1.8.0=pypi_0
notebook=6.5.4=pypi_0
notebook-shim=0.2.3=pypi_0
numba=0.49.0=pypi_0
numpy=1.19.5=pypi_0
nuscenes-devkit=1.1.10=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
open-clip-torch=2.16.0=pypi_0
openai=1.1.1=pypi_0
opencv-python=4.7.0.72=pypi_0
opencv-python-headless=4.5.5.64=pypi_0
opendatasets=0.1.22=pypi_0
openexr=1.3.9=pypi_0
openmim=0.3.7=pypi_0
openssl=3.0.11=h7f8727e_2
ordered-set=4.1.0=pypi_0
packaging=23.1=pypi_0
pandas=1.4.0=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.1=pypi_0
pathtools=0.1.2=pypi_0
pathy=0.10.2=pypi_0
pccm=0.4.8=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=pypi_0
pip=23.2.1=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.11.0=pypi_0
plotly=5.13.1=pypi_0
pluggy=1.0.0=pypi_0
plyfile=0.9=pypi_0
portalocker=2.7.0=pypi_0
pre-commit=3.4.0=pypi_0
preshed=3.0.9=pypi_0
prettytable=3.7.0=pypi_0
prometheus-client=0.16.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=10.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pybind11=2.11.1=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.6=pypi_0
pycodestyle=2.9.1=pypi_0
pycparser=2.21=pypi_0
pydantic=2.4.2=pypi_0
pydantic-core=2.10.1=pypi_0
pydeck=0.8.1b0=pypi_0
pyflakes=2.5.0=pypi_0
pygments=2.15.1=pypi_0
pyparsing=3.0.9=pypi_0
pyquaternion=0.9.9=pypi_0
pyrsistent=0.19.3=pypi_0
pytest=7.3.1=pypi_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
python-magic=0.4.27=pypi_0
python-slugify=8.0.1=pypi_0
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=pypi_0
qtconsole=5.4.3=pypi_0
qtpy=2.3.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.5.5=pypi_0
requests=2.30.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.3.5=pypi_0
rsa=4.9=pypi_0
salesforce-lavis=1.0.2=pypi_0
sanmodule=0.1=dev_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.9.1=pypi_0
send2trash=1.8.2=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=1.22.1=pypi_0
setuptools=67.6.0=pypi_0
shapely=1.8.5=pypi_0
simplejson=3.19.1=pypi_0
six=1.16.0=pypi_0
smart-open=6.4.0=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
spacy=3.7.1=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
spconv=2.3.6=pypi_0
sqlite=3.41.2=h5eee18b_0
srsly=2.4.8=pypi_0
stack-data=0.6.2=pypi_0
streamlit=1.27.2=pypi_0
submitit=1.4.5=pypi_0
svgwrite=1.4.3=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.2.2=pypi_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=pypi_0
terminaltables=3.1.10=pypi_0
text-unidecode=1.3=pypi_0
thinc=8.2.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.4.12=pypi_0
timm=0.4.12=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
torch=1.10.0+cu113=pypi_0
torchaudio=0.10.0+cu113=pypi_0
torchvision=0.11.1+cu113=pypi_0
tornado=6.3.1=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.26.1=pypi_0
tree=0.2.4=pypi_0
trimesh=2.35.39=pypi_0
typeguard=2.13.3=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.8.0=pypi_0
tzlocal=5.1=pypi_0
uri-template=1.2.0=pypi_0
urllib3=1.26.15=pypi_0
validators=0.22.0=pypi_0
virtualenv=20.24.5=pypi_0
wandb=0.15.2=pypi_0
wasabi=1.1.2=pypi_0
watchdog=3.0.0=pypi_0
wcwidth=0.2.6=pypi_0
weasel=0.3.2=pypi_0
webcolors=1.13=pypi_0
webdataset=0.2.57=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=2.3.3=pypi_0
wheel=0.41.2=py38h06a4308_0
widgetsnbextension=4.0.7=pypi_0
wrapt=1.15.0=pypi_0
xz=5.4.2=h5eee18b_0
yacs=0.1.8=pypi_0
yapf=0.33.0=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
